:art: Refactor Plan package to follow Clean Architecture
