+++ src/main/java/com/trip/IronBird_Server/plan/adapter/controller/PlanController.java
@@ -2,7 +2,7 @@ package com.trip.IronBird_Server.plan.adapter.controller;
 
 import com.trip.IronBird_Server.common.custom.CustomUserDetails;
 import com.trip.IronBird_Server.plan.adapter.dto.PlanDto;
-import com.trip.IronBird_Server.plan.application.service.PlanService;
+import com.trip.IronBird_Server.plan.application.service.PlanServiceImp;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,7 +16,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PlanController {
 
-    private final PlanService planService;
+    private final PlanServiceImp planServiceImp;
 
     /**
      @
@@ -25,7 +25,7 @@ public class PlanController {
     @GetMapping
     public List<PlanDto> getAllPlans(){
 
-        return planService.getAllPlans();
+        return planServiceImp.getAllPlans();
     }
 
     /**
@@ -35,7 +35,7 @@ public class PlanController {
     @GetMapping("/user/{userId}")
     public List<PlanDto> getPlansByUserId(@PathVariable("userId") Long userId) {    //PathVariable 에 명시적으로 userId 를 등록하여 경로 변수에 인식
 
-        return planService.getPlansByUserId(userId);
+        return planServiceImp.getPlansByUserId(userId);
     }
 
     /**
@@ -47,7 +47,7 @@ public class PlanController {
                                         @AuthenticationPrincipal CustomUserDetails userDetails){
 
         Long userIdFromToken = userDetails.getId();
-        PlanDto createdPlan = planService.createPlan(planDto,userIdFromToken);
+        PlanDto createdPlan = planServiceImp.createPlan(planDto,userIdFromToken);
         return ResponseEntity.status(HttpStatus.CREATED).body(createdPlan);
     }
 
@@ -59,7 +59,7 @@ public class PlanController {
     @PutMapping("/update/{planId}")
     public ResponseEntity<PlanDto> updatePlanById(@PathVariable("planId") Long planId,
                                                   @RequestBody PlanDto planDto){
-        PlanDto updatePlan = planService.updatePlan(planId, planDto);
+        PlanDto updatePlan = planServiceImp.updatePlan(planId, planDto);
         return ResponseEntity.status(HttpStatus.CREATED).body(updatePlan);
     }
 
@@ -70,7 +70,7 @@ public class PlanController {
     @DeleteMapping("/{planId}")
     public ResponseEntity<String> deletePlan(@PathVariable("planId") Long PlanId){
         try {
-            planService.deletePlan(PlanId);
+            planServiceImp.deletePlan(PlanId);
 
             return ResponseEntity.ok("플랜이 삭제되었습니다.");
         }catch (Exception e){
