+++ src/main/java/com/trip/IronBird_Server/plan/domain/Plan.java
@@ -14,8 +14,6 @@ import java.time.LocalDateTime;
 
 @Data
 @NoArgsConstructor
-@AllArgsConstructor
-@Builder
 @Entity
 public class Plan {
 
@@ -42,4 +40,18 @@ public class Plan {
     private LocalDateTime modified_time;
 
 
+    @Builder
+    public Plan(User user, String startedDate, String endDate, LocalDateTime created_time, LocalDateTime modified_time) {
+
+        this.user = user;
+        this.startedDate = startedDate;
+        this.endDate = endDate;
+        this.created_time = created_time;
+        this.modified_time = modified_time;
+    }
+    public void updateDates(String startedDate, String endDate) {
+        this.startedDate = startedDate;
+        this.endDate = endDate;
+        this.modified_time = LocalDateTime.now();
+    }
 }
